Change a lot of the Mesh routines to signed int
Unsigned ints are dangerous because they dont work with Blitz arrays. Easier to
just not have them floating around in the first place

+++ src/core/mesh.cpp
@@ -518,7 +518,7 @@ namespace mocc {
         return 0;
     }
 
-    size_t Mesh::coarse_boundary_cell( Point2 p, int octant ) const {
+    int Mesh::coarse_boundary_cell( Point2 p, int octant ) const {
         assert( octant < 5 );
 
         bool on_x = false;
